Remove other components from App.js
