+++ app/frontend/src/utilities.js
@@ -25,3 +25,48 @@ export const makeId = (ty, value) => {
 	const idString = JSON.stringify({ type: ty, value: id })
 	return idString
 }
+
+const findRef = (refs, ids) =>
+	refs.find(ref => ref.names.find(v => ids.includes(v.id)))
+
+const findId = (state, id) => state.map[id]
+
+export const formatQualifier = (qualifier, state) => {
+	if (qualifier === undefined) return ''
+
+	const idObject = parseId(qualifier.id)
+	if (idObject.type !== 'qualifier' && idObject.type !== 'db_qualifier')
+		throw new Error(
+			`Object with id ${qualifier.id} is not a structured name.`
+		)
+
+	// This distinction is necessary since the wizard assumes defining a new qualifier includes
+	// defining a new name so the name is stored directly in the qualifier
+	const qualifierName =
+		idObject.type === 'db_qualifier'
+			? findId(state, qualifier.qualifier_name_id)
+			: qualifier
+
+	return qualifierName ? qualifierName.name : ''
+}
+
+export const formatStructuredName = (structuredName, state) => {
+	const idObject = parseId(structuredName.id)
+	if (
+		idObject.type !== 'structured_name' &&
+		idObject.type !== 'db_structured_name'
+	)
+		throw new Error(
+			`Object with id ${structuredName.id} is not a structured name.`
+		)
+
+	const name = findId(state, structuredName.name_id)
+	const qualifierName = formatQualifier(
+		findId(state, structuredName.qualifier_id),
+		state
+	)
+	const location = findId(state, structuredName.location_id)
+	return `${name ? name.name : ''} / ${qualifierName} / ${
+		location ? location.name : ''
+	}`
+}
